requirements.txt
pyspark==3.1.2
pandas==1.4.3
pyyaml==6.0